requirements.txt
Package             Version
------------------- -------
asgiref             3.2.10
astroid             2.4.2
autopep8            1.5.4
dj-database-url     0.5.0
Django              3.1.2
django-cors-headers 3.5.0
django-heroku       0.3.1
djangorestframework 3.12.1
gunicorn            20.0.4
isort               5.6.4
lazy-object-proxy   1.4.3
mccabe              0.6.1
pip                 20.2.3
psycopg2            2.8.6
pycodestyle         2.6.0
pylint              2.6.0
pytz                2020.1
setuptools          50.3.0
six                 1.15.0
sqlparse            0.4.1
toml                0.10.1
wheel               0.35.1
whitenoise          5.2.0
wrapt               1.12.1
